// packages/decoupled-components/src/types.ts
export type commonTypes = {
  variant?: 'flat' | 'solid' | 'outlined' | 'soft'
  color?:
    | 'primary'
    | 'secondary'
    | 'danger'
    | 'warning'
    | 'success'
    | 'info'
    | 'default'
    | 'active'
    | 'light'
    | 'dark'
  size?: 'sm' | 'base' | 'lg' | 'xl'
  radius?: 'none' | 'soft' | 'round'
}